user_service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
 "pydantic>=2.11.7",
 "result>=0.17.0",
]

[dependency-groups]
dev = ["datamodel-code-generator>=0.31.0", "ruff>=0.11.13"]
